added error printing using println(will be replaced by some proper more correct error handling, but will do for debugging).

+++ src/real/Objects/ConditionOperations/Atomic/AttributeLiteral.java
@@ -24,9 +24,9 @@ public class AttributeLiteral extends UnaryConditionBase
     }
 
     @Override
-    public int evaluateNumber(Row row)
+    public float evaluateNumber(Row row)
     {
-        return Integer.parseInt(row.getValue(attributeName));
+        return Float.parseFloat(row.getValue(attributeName));
     }
 
     @Override
