+++ src/real/Objects/ConditionOperations/Atomic/StringLiteral.java
@@ -22,7 +22,7 @@ public class StringLiteral extends UnaryConditionBase
     }
 
     @Override
-    public int evaluateNumber(Row row)
+    public float evaluateNumber(Row row)
     {
         throw new UnsupportedOperationException("Not supported yet.");
     }
